+++ tests/support/decorator_order.py
@@ -1,0 +1,13 @@
+from fabric.api import task, hosts, roles
+
+
+@hosts('whatever')
+@task
+def foo():
+    pass
+
+# There must be at least one unmolested new-style task for the decorator order
+# problem to appear.
+@task
+def caller():
+    pass